Add files via upload
The above files contain C, Assembly and machine code for a program that prints out reversible prime squares.
Theoretical Background
Both 169 and 961 are the square of a prime. 169 is the reverse of 961.
We call a number a reversible prime square if:
• It is not a palindrome, and
• It is the square of a prime, and Its reverse is also the square of a prime.
• 169 and 961 are not palindromes, so both are reversible prime squares


To run the C code, use C++ compiler or Microsoft Visual Studio.
To run the Assembly code, use QtSpim or Mars compiler.